Nueva estructura de carpeta y elementos de la aplicación
