+++ app/src/main/java/com/example/mynotes/data/DataSource.kt
@@ -1,0 +1,2 @@
+package com.example.mynotes.data
+
