+++ app/src/main/java/com/example/mynotes/ui/StartNotesScreen.kt
@@ -1,0 +1,2 @@
+package com.example.mynotes.ui
+
